feat: in-prog adlister exercises

+++ src/main/java/ListAdsDao.java
@@ -18,7 +18,7 @@ public class ListAdsDao implements Ads {
         }
         // we'll assign an "id" here based on the size of the ads list
         // really the database would handle this
-        ad.setId((long) ads.size());
+        ad.setId(ads.size());
         ads.add(ad);
         return ad.getId();
     }
